requirements.txt
fastapi==0.128.0
google-api-core==2.28.1
google-auth==2.45.0
google-cloud-core==2.5.0
google-cloud-storage==3.7.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
httpx==0.28.1
httpx-sse==0.4.3
litellm==1.80.11
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pypdf==6.7.3
python-dotenv==1.2.1
python-multipart==0.0.21
strands-agents==1.27.0
uvicorn==0.40.0
